fix(lint): split manager_create.go under the 1000-line limit
#2719 pushed daemon/manager_create.go to 1011 lines and turned master's
file-length lint red. I merged that PR, so this is mine to fix.

Extracts the archived-title and held-branch reuse helpers into
daemon/manager_create_titles.go — a coherent group that runs under the manager
lock and answers one question between them: can this title be used, and if not,
what happens to the archived record holding it. Pure move plus the imports each
file now needs; no behaviour change.

  manager_create.go        1011 -> 524
  manager_create_titles.go         502

Not grandfathered in scripts/file-length-allowlist.txt — the limit exists to
force exactly this split, and #1145 says an allowlist entry is for when a split
is genuinely infeasible, which it was not here.

Verified: go build ./... clean, go vet ./daemon/ clean, daemon test package
compiles, gofmt -l clean, scripts/lint-file-length.sh passes (1139 files, 0
grandfathered).
